+++ tests/cli/config/test_show.py
@@ -13,6 +13,7 @@ def test_default_scrubbed(ddqa, config_file, helpers):
         """
         repo = ""
         cache_dir = ""
+        include_labels = []
 
         [github]
         user = "foo"
@@ -38,6 +39,7 @@ def test_reveal(ddqa, config_file, helpers):
         """
         repo = ""
         cache_dir = ""
+        include_labels = []
 
         [github]
         user = "foo"